requirements.txt
lxml_html_clean
streamlit
fastapi
uvicorn
google-generativeai
newspaper3k
beautifulsoup4
requests